Correct bug: domain class was using id of type Integer while database was using id of type Long.

+++ src/main/java/com/airpnp/data/ParkingSpaceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParkingSpaceRepository extends JpaRepository<ParkingSpace, Long> {
+public interface ParkingSpaceRepository extends JpaRepository<ParkingSpace, Integer> {
 }
